refactor ui api client helpers
